Support loading hosted PGNs
PGN files can now be hosted on the server. Available files are listed in the
Sidebar and can be downloaded and played on demand. Some small example PGNs
are included as well.

For now, all hosted PGN files must be enumerated in `src/pngList.ts`.

User retains the option of loading a local PGN as well.