Move "Load PGN" and "About" buttons to be part of Chessboard.vue component
The "Load PGN" and "About" buttons previously were under the "Home.vue"
view. Moving them under Chessboard.vue makes it easier to develop
component layout decisions with a clear left/right structure, as opposed to
having to account for a separate button component just under the page header.

With this change, the top of the board header now aligns with the bottom
of the page header.