Set initial chessboard size to be relative to viewport size
The initial chessboard size will now be set relative to the viewport size when
below threshold of 600px. This provides a better experience on mobile browsers
as the chessboard size will scale down with available screen space.

When below this threshold, the Game List will change from a select list-box
into a dropdown menu to better accommodate the reduced screen space.