Give move annotations better formatting and structure
In the initial project stages, move annotations were simply concatenated
and reduced to a string in order to provide a quick way to access puzzle
answers. While functional, this made annotations difficult to read and
difficult to programmatically extend to include annotation RAVs
("Recursive Annotation Variations") which can provide additional context
to the answers. RAVs are not yet presented to the user, but this work
makes the future implementation significantly easier. As RAVs are frequently
included in move annotations and are a recursive data type, it was important
to get the data structure correct before exploring visual changes.

This commit gives better structure to the move annotation model. No style
changes have been applied beyond organizing the move annotations into a
column, but it is still an improvement over the concatenated string.

Included is the evaluation of Numeric Annotation Glyphs ("NAGs") in move
annotations which permits additional commentary in the form of traditional
chess symbols (e.g. "!" = "Good Move") as well as more expressive commentary
as specified in the PGN standard. (e.g. "White has a moderately weak
pawn structure.").