Add default "Select a game/collection" option to prevent unselected state on mobile
When a default value is not provided for a `form-select` component, on mobile
browsers it will default to an "unselected" state. When this happens, it becomes
unintuitive to select the first available option because it visually appears already
selected. The workaround for the user is to select a different option before selecting
the first option. This is a poor experience.

The bootstrap-vue recommended solution is to provide a disabled option with an
empty value as the first option. This commit implements this solution.